requirements.txt
discord
python-dotenv
python-Levenshtein
wikitextparser
